requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
billiard==4.2.3
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.10
dvc==3.64.0
dvc-data==3.16.12
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.20.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grandalf==0.8
gto==1.9.0
httplib2==0.31.0
hydra-core==1.3.2
idna==3.11
iterative-telemetry==0.0.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
oauth2client==4.1.3
omegaconf==2.3.0
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.5.3
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.228
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.14
yarl==1.22.0
zc.lockfile==4.0
